Reorganize graspcli
