+++ .flake8
@@ -5,4 +5,5 @@ max-complexity = 10
 # We import modules here to make the API cleaner
 per-file-ignores = 
   **/fetch_api/__init__.py:F401
+  **/uw_manipulation/__init__.py:F401
   **/joint_state_reader/__init__.py:F401
