+++ fetch_api/src/fetch_api/__init__.py
@@ -6,4 +6,3 @@ from .gripper_viz import gripper_interactive_marker
 from .head import Head
 from .moveit_goal_builder import MoveItGoalBuilder
 from .torso import Torso
-from .graspcli import GraspingClient
