+++ uw_manipulation/src/uw_manipulation/__init__.py
@@ -1,0 +1,1 @@
+from .graspcli import GraspingClient
