+++ uw_manipulation/src/uw_manipulation/graspcli.py
@@ -2,14 +2,14 @@ import copy
 import actionlib
 import rospy
 
-from math import sin, cos
 from moveit_python import (MoveGroupInterface,
                            PlanningSceneInterface,
                            PickPlaceInterface)
 from moveit_python.geometry import rotate_pose_msg_by_euler_angles
 from grasping_msgs.msg import FindGraspableObjectsAction, FindGraspableObjectsGoal
 # from geometry_msgs.msg import PoseStamped
-from moveit_msgs.msg import PlaceLocation, MoveItErrorCodes
+from moveit_msgs.msg import PlaceLocation
+
 
 # Tools for grasping
 class GraspingClient(object):
@@ -45,7 +45,7 @@ class GraspingClient(object):
         idx = -1
         for obj in find_result.objects:
             idx += 1
-            obj.object.name = "object%d"%idx
+            obj.object.name = "object%d" % idx
             self.scene.addSolidPrimitive(obj.object.name,
                                          obj.object.primitives[0],
                                          obj.object.primitive_poses[0],
@@ -72,7 +72,7 @@ class GraspingClient(object):
         self.surfaces = find_result.support_surfaces
 
     def getGraspableCube(self):
-        graspable = None
+        # graspable = None
         for obj in self.objects:
             # self.getCokeReference(obj)
             # need grasps
@@ -130,32 +130,23 @@ class GraspingClient(object):
 
     def place(self, block, pose_stamped):
         places = list()
-        l = PlaceLocation()
-        l.place_pose.pose = pose_stamped.pose
-        l.place_pose.header.frame_id = pose_stamped.header.frame_id
+        location = PlaceLocation()
+        location.place_pose.pose = pose_stamped.pose
+        location.place_pose.header.frame_id = pose_stamped.header.frame_id
 
         # copy the posture, approach and retreat from the grasp used
-        l.post_place_posture = self.pick_result.grasp.pre_grasp_posture
-        l.pre_place_approach = self.pick_result.grasp.pre_grasp_approach
-        l.post_place_retreat = self.pick_result.grasp.post_grasp_retreat
-        places.append(copy.deepcopy(l))
+        location.post_place_posture = self.pick_result.grasp.pre_grasp_posture
+        location.pre_place_approach = self.pick_result.grasp.pre_grasp_approach
+        location.post_place_retreat = self.pick_result.grasp.post_grasp_retreat
+        places.append(copy.deepcopy(location))
         # create another several places, rotate each by 360/m degrees in yaw direction
-        m = 16 # number of possible place poses
+        m = 16  # number of possible place poses
         pi = 3.141592653589
         for i in range(0, m-1):
-            l.place_pose.pose = rotate_pose_msg_by_euler_angles(l.place_pose.pose, 0, 0, 2 * pi / m)
-            places.append(copy.deepcopy(l))
+            location.place_pose.pose = rotate_pose_msg_by_euler_angles(location.place_pose.pose, 0, 0, 2 * pi / m)
+            places.append(copy.deepcopy(location))
 
         success, place_result = self.pickplace.place_with_retry(block.name,
                                                                 places,
                                                                 scene=self.scene)
         return success
-
-    def tuck(self):
-        joints = ["shoulder_pan_joint", "shoulder_lift_joint", "upperarm_roll_joint",
-                  "elbow_flex_joint", "forearm_roll_joint", "wrist_flex_joint", "wrist_roll_joint"]
-        pose = [1.32, 1.40, -0.2, 1.72, 0.0, 1.66, 0.0]
-        while not rospy.is_shutdown():
-            result = self.move_group.moveToJointPosition(joints, pose, 0.02)
-            if result.error_code.val == MoveItErrorCodes.SUCCESS:
-                return
